relocate calculation constants, update tanks parser
