+++ bio_diversity/data_parsers/calibration.py
@@ -1,10 +1,5 @@
-from django.core.exceptions import ValidationError
-from django.db import IntegrityError
-import pandas as pd
-
 from bio_diversity import models
 from bio_diversity import utils
-from bio_diversity.static import calculation_constants
 from bio_diversity.utils import DataParser
