+++ bio_diversity/data_parsers/sites.py
@@ -1,12 +1,9 @@
 
 from django.core.exceptions import ValidationError
 from django.db import IntegrityError
-from django.db.models import Count as django_Count
-from django.db.models import Q
 
 from bio_diversity import models
 from bio_diversity import utils
-from bio_diversity.static import calculation_constants
 from bio_diversity.utils import DataParser
