+++ bio_diversity/data_parsers/spawning.py
@@ -3,10 +3,9 @@ from django.core.exceptions import ValidationError
 from django.db import IntegrityError
 
 
-from bio_diversity import models
+from bio_diversity import models, calculation_constants
 from bio_diversity import utils
-from bio_diversity.static import calculation_constants
-from bio_diversity.static.calculation_constants import prio_dict
+from bio_diversity.calculation_constants import prio_dict
 from bio_diversity.utils import DataParser
