+++ bio_diversity/data_parsers/tagging.py
@@ -1,11 +1,9 @@
 from django.core.exceptions import ValidationError
 from django.db import IntegrityError
-import pandas as pd
 from django.db.models.functions import Length
 
-from bio_diversity import models
+from bio_diversity import models, calculation_constants
 from bio_diversity import utils
-from bio_diversity.static import calculation_constants
 from bio_diversity.utils import DataParser
