+++ bio_diversity/test/BioFactoryFloor.py
@@ -1,11 +1,9 @@
 
 import factory
 import pytz
-from django.utils import timezone
 from faker import Factory
 
-from bio_diversity import models
-from bio_diversity.static import calculation_constants
+from bio_diversity import models, calculation_constants
 
 faker = Factory.create()
 
@@ -2113,8 +2111,11 @@ class RelcFactory(factory.django.DjangoModelFactory):
     subr_id = factory.SubFactory("bio_diversity.test.BioFactoryFloor.SubrFactory")
     min_lat = factory.lazy_attribute(lambda o: faker.random_int(0, 45))
     max_lat = factory.lazy_attribute(lambda o: faker.random_int(50, 90))
-    min_lon = factory.lazy_attribute(lambda o: faker.random_int(calculation_constants.min_long, (calculation_constants.min_long + calculation_constants.max_long) / 2))
-    max_lon = factory.lazy_attribute(lambda o: faker.random_int((calculation_constants.min_long + calculation_constants.max_long) / 2 + 5, calculation_constants.max_long))
+    min_lon = factory.lazy_attribute(lambda o: faker.random_int(calculation_constants.min_long, (
+                calculation_constants.min_long + calculation_constants.max_long) / 2))
+    max_lon = factory.lazy_attribute(lambda o: faker.random_int((calculation_constants.min_long +
+                                                                 calculation_constants.max_long) / 2 + 5,
+                                                                calculation_constants.max_long))
 
     created_by = factory.lazy_attribute(lambda o: faker.name())
     created_date = factory.lazy_attribute(lambda o: faker.date())
